feat(save product): create SaveButton and pass product changes to Button for save
We use useEffect to pass changes up to the screen, and andother useEffect to pass those changes to
navgation so the data is in the Save Buttton component. SaveButton can then access redux and fire
save actions to complete the process. BUT... this seems long winded... to PASS data up to a menu
button, as opposed to the menu button passig back down.. There is a better away... FIND IT!!!