feat(creat-product|getproducts): add a CREATE_PRODUCT action for new products. get products on init
much wrk here around use of fetch and especially error handling , which is unclear. For xample, I
cant get to the respnse code via the response object and the response data is unclear if its null or
undefined. Anyway, the slower development and additional failure handling code may be a reason to
switch to axios?  Lastly, there is adebate on how to return, (a) assume success and return data,
plus throw errors on failure, or (b) always return a status object that contains { status, error |
data }. Need t standardise on this.