fix(not saving new produc): fix error where refresh of useCallback function did not call useEffect
problem is good example of useEffect dependencies. the useEffect ONLY worked if it was located AFTER
the useCallback function , so in effect it got the recompiled version simply by order of code
exacution, but did NOT get called by change to the memoized (useCallback) function. Problem also
existed wit useMemo() so not useCallback specific. A HACK fix was to force the useCallback()
dependencies ALSO to the eseEffect(). This worked but a future progremmer coudl easily have removd
those dependencies, seeing no rational need for them. (In fact, that was what I had implicitly done
on writing the code.  And withuseCallback it *should* be sufficient to depend on that. Anyway....
the BEST (and recommended) solution was to drop useCallback() and just define the function INSIDE
useEffect() and hence put the original dependencies on useEffect. Phew!